Move all step definitions to one single file, and remove some redundant codes
