+++ integ-tests/src/cucumberTest/java/cucumber/steps/wrappers/DataServiceWrapper.java
@@ -26,9 +26,7 @@ import com.amazonaws.blox.dataservicemodel.v1.model.wrappers.UpdateEnvironmentRe
 import com.amazonaws.blox.dataservicemodel.v1.model.wrappers.DeleteEnvironmentResponse;
 import java.util.function.Consumer;
 
-import cucumber.steps.helpers.ThrowingFunction;
 import lombok.RequiredArgsConstructor;
-import org.apache.commons.lang3.Validate;
 
 @RequiredArgsConstructor
 public class DataServiceWrapper extends MemoizedWrapper {
@@ -43,17 +41,17 @@ public class DataServiceWrapper extends MemoizedWrapper {
 
   public DescribeEnvironmentResponse describeEnvironment(
       DescribeEnvironmentRequest describeEnvironmentRequest) {
-    return memoizeInputAndCall(describeEnvironmentRequest, dataService::describeEnvironment);
+    return memoizeFunction(describeEnvironmentRequest, dataService::describeEnvironment);
   }
 
   public UpdateEnvironmentResponse updateEnvironment(
       UpdateEnvironmentRequest updateEnvironmentRequest) {
-    return memoizeInputAndCall(updateEnvironmentRequest, dataService::updateEnvironment);
+    return memoizeFunction(updateEnvironmentRequest, dataService::updateEnvironment);
   }
 
   public DeleteEnvironmentResponse deleteEnvironment(
       DeleteEnvironmentRequest deleteEnvironmentRequest) {
-    return memoizeInputAndCall(deleteEnvironmentRequest, dataService::deleteEnvironment);
+    return memoizeFunction(deleteEnvironmentRequest, dataService::deleteEnvironment);
   }
 
   public void tryCreateEnvironment(CreateEnvironmentRequest createEnvironmentRequest) {
@@ -70,14 +68,6 @@ public class DataServiceWrapper extends MemoizedWrapper {
 
   public void tryDeleteEnvironment(DeleteEnvironmentRequest deleteEnvironmentRequest) {
     captureException(deleteEnvironmentRequest, this::deleteEnvironment);
-  }
-
-  private <T, R> R memoizeInputAndCall(final T input, ThrowingFunction<T, R> fn) {
-    Validate.notNull(input);
-    addToHistory((Class<T>) input.getClass(), input);
-    final R response = fn.apply(input);
-    addToHistory((Class<R>) response.getClass(), response);
-    return response;
   }
 
   private <T> void captureException(final T input, final Consumer<T> consumer) {
